requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
sentence-transformers==2.2.2
anthropic==0.7.0
pydantic==2.12.5
typing_extensions==4.15.0
numpy>=1.26.0
python-dotenv==1.0.0
tensorflow==2.20.0
scikit-learn==1.5.0
pandas==2.2.3
click==8.1.7